Homework 7 web

+++ goldbuy/src/main/java/com/senla/haltvinizki/dao/CredentialsDao.java
@@ -1,9 +1,9 @@
 package com.senla.haltvinizki.dao;
 
-import com.senla.haltvinizki.entity.credentials.Credentials;
+import com.senla.haltvinizki.entity.Credentials;
 
-public interface CredentialsDao extends GenericDao<Credentials,Integer> {
+public interface CredentialsDao extends GenericDao<Credentials,Long> {
 
-    Credentials getCredentialsWithUser(int id);
+    Credentials getCredentialsWithUser(Long id);
 
 }
