+++ goldbuy/src/main/java/com/senla/haltvinizki/dao/ProductConfigurationDao.java
@@ -1,8 +1,8 @@
 package com.senla.haltvinizki.dao;
 
-import com.senla.haltvinizki.entity.productCofniguration.ProductConfiguration;
+import com.senla.haltvinizki.entity.ProductConfiguration;
 
-public interface ProductConfigurationDao extends GenericDao<ProductConfiguration, Integer> {
-    ProductConfiguration getProductConfigWithProduct(int id);
+public interface ProductConfigurationDao extends GenericDao<ProductConfiguration, Long> {
+    ProductConfiguration getProductConfigWithProduct(Long id);
 
 }
